Java language level migration: --Using  java7 Objects.equals where possible --Repacing StringBuffer with StringBuilder --Removed unnecessary boxing and unboxing --Replacing while loop with forEach where applicable --Merging identical try blocks into single block --Replaced for loop with forEach where possible

+++ wavemaker-app-runtime-core/src/main/java/com/wavemaker/runtime/security/rememberme/WMAppRememberMeServices.java
@@ -74,7 +74,7 @@ public class WMAppRememberMeServices extends AbstractWMRememberMeServices {
         long tokenExpiryTime;
 
         try {
-            tokenExpiryTime = new Long(cookieTokens[1]).longValue();
+            tokenExpiryTime = new Long(cookieTokens[1]);
         } catch (NumberFormatException nfe) {
             throw new InvalidCookieException("Cookie token[1] did not contain a valid number (contained '" +
                     cookieTokens[1] + "')");
